koji-ssl-admin: write a ".chain.crt" file when signing server certs
After the user signs an HTTP server certificate, the user must
concatenate the HTTP server public cert and the CA public cert together
so that the web server can send the full certificate chain.

Perform this concatenation as a part of the "koji-ssl-admin sign"
sub-command.

The purpose of this change is to reduce the manual steps a user must
perform to set up an SSL certificate suitable for Apache.