koji-ssl-admin: use re.sub to determine filenames
Prior to this change, koji-ssl-admin would determine some file paths by
globally searching-and-replacing the strings ".csr" or ".crt".

This would cause problems if the Koji server hostname or Koji username
contained the strings ".crt" or ".csr".

Use a regular expression to determine the correct filename instead, by
only replacing the *final* occurrences of ".csr" or ".crt".