koji-search-containers: raise when metadata.json download fails
Prior to this change, if the Koji packages web server returned an HTTP
error for the container build metadata.json file, we would not catch
this error. The program might crash later on when parsing
response.json(), or elsewhere.

Raise immediately if we get an HTTP error when downloading the
metadata.json file. This makes it easier for developers to understand
the root cause of a failure in this code.