koji-ssl-admin: use re.sub for pkcs12 filename
Prior to this change, koji-ssl-admin would determine the path for the
pkcs12 bundle file by globally searching-and-replacing the string ".key"
in the key filename.

This would cause problems if the Koji server hostname or Koji username
contained the string ".key".

Use a regular expression to determine the correct filename instead, by
only replacing the *final* occurrences of ".key".

This fix is similar to the change in
b0d83f317ea38845dc7b345ba5896ac76bc5cde9.